+++ src/routes/characters/+page.svelte
@@ -4,29 +4,16 @@
     import { CharactersDB, locales } from '@db/Database';
     import MarkupHtmlView from '@components/concepts/MarkupHTMLView.svelte';
     import { getUser } from '@components/project/Contexts';
-    import Button from '@components/widgets/Button.svelte';
     import { firestore } from '@db/firebase';
     import Feedback from '@components/app/Feedback.svelte';
-    import { goto } from '$app/navigation';
-    import Spinning from '@components/app/Spinning.svelte';
     import {
         characterToSVG,
         type Character,
     } from '../../db/characters/Character';
     import Link from '@components/app/Link.svelte';
+    import NewCharacterButton from './NewCharacterButton.svelte';
 
     const user = getUser();
-
-    let creating: boolean | undefined = $state(false);
-
-    async function addCharacter() {
-        creating = true;
-        const id = await CharactersDB.createCharacter();
-        if (id) {
-            creating = false;
-            goto(`/character/${id}`);
-        } else creating = undefined;
-    }
 
     let characters = $derived(CharactersDB.getEditableCharacters());
 </script>
@@ -80,23 +67,7 @@
             >{$locales.get((l) => l.ui.page.characters.error.noauth)}</Feedback
         >
     {:else}
-        {#if creating}
-            <Spinning></Spinning>
-        {:else if creating === undefined}
-            <Feedback
-                >{$locales.get(
-                    (l) => l.ui.page.characters.error.create,
-                )}</Feedback
-            >
-        {:else}
-            <Button
-                tip={$locales.get((l) => l.ui.page.characters.button.new)}
-                action={addCharacter}
-                active={!creating}
-                large
-                icon="+"
-            ></Button>
-        {/if}
+        <NewCharacterButton></NewCharacterButton>
 
         <div class="characters">
             {#each characters.values() as character}